+++ changelog.md
@@ -3,6 +3,8 @@
 ### 1.0.RC4 ###
 
 - [Issue 140](https://github.com/stormpath/stormpath-sdk-java/issues/140): Verification token can now be obtained via the `EmailVerificationToken#getToken()` method.
+- [Issue 150](https://github.com/stormpath/stormpath-sdk-java/issues/150): Added support for the `PasswordPolicy` resource in Directories.
+- [Issue 151](https://github.com/stormpath/stormpath-sdk-java/issues/151): Fixed issue where `UserAgent` was not being able to read POM properties from Virtual File System.
 
 ### 1.0.RC3.1 ###